Update binding metadata structure
